Making cleaner main file and making screen a class
